Add changes requested for modularisation
- rename the model memory to “My Model Library”
- change "titanic model information" to "titanic task information"
- add Node version to the readme
- naming convention update (snake case)
- folder name (no space, respect the snake case convention)
- remove peers server from readme & add the new peer credentials (test with gcloud, martin m has set up credentials already)
