+++ mobile-browser-based-version/src/components/building_frames/ImageUploadFrame.vue
