--- mobile-browser-based-version/src/helpers/Data Validation Script/helpers-image-tasks.js
@@ -1,3 +1,0 @@
-export async function check_data(training_data){
-    return {accepted: Object.keys(training_data).length>1}
-}
