+++ mobile-browser-based-version/src/helpers/data_validation_script/helpers-image-tasks.js
@@ -1,0 +1,3 @@
+export async function checkData(trainingData){
+    return {accepted: Object.keys(trainingData).length>1}
+}
